+++ src/modules/Button/Button.tsx
@@ -1,50 +1,53 @@
 import React from 'react';
 import styled, { css } from 'styled-components';
-import { borderRadius, calcSize, Colours, trainsitionTime } from '../../mainStyles';
+import { borderRadius, calcFontSize, Colours, trainsitionTime } from '../../mainStyles';
 import LoadingAnimation from '../Loader/LoadingAnimation';
 
 export interface IButtonProps {
   className?: string;
+  error?: boolean;
   primary?: boolean;
   link?: boolean;
   disabled?: boolean;
-  value?: string;
   onClick?: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
-  id?: string;
   name?: string;
   type?: 'submit' | 'reset' | 'button';
-  text?: string;
   size?: 'xs' | 'sm' | 'md' | 'lg';
-  error?: boolean;
-  iconLeft?: React.ReactElement;
-  iconRight?: React.ReactElement;
+  icon?: React.ReactElement;
   isLoading?: boolean;
-  bordered?: boolean;
+  children?: string;
+  hasBorder?: boolean;
 }
 
-const SText = styled.div<{ hasIconLeft: boolean; hasIconRight: boolean }>`
-  ${({ hasIconLeft }) => hasIconLeft && `margin-left: 13px;`}
-  ${({ hasIconRight }) => hasIconRight && ` margin-right: 13px;`}
-`;
-
-export const ButtonWrapper = styled.button<IButtonProps>`
+export const ButtonWrapper = styled.button<Omit<IButtonProps, 'children'>>`
   transition: all ${trainsitionTime};
   border-radius: ${borderRadius};
   white-space: nowrap;
   cursor: pointer;
   background-color: transparent;
   font-size: inherit;
-  padding: 4px 10px;
-  border: 1px solid ${Colours.grey};
+  border: none;
 
-  ${({ bordered }) =>
-    !bordered &&
+  padding: 0;
+  display: flex;
+  align-items: center;
+
+  ${({ hasBorder }) =>
+    hasBorder &&
     css`
-      border: none;
+      border: 1px solid ${Colours.grey};
     `}
 
   &:focus {
     outline: none;
+  }
+
+  & > span > svg {
+    font-size: 80%;
+  }
+
+  & > span + span {
+    padding-left: 0.5rem;
   }
 
   ${({ disabled }) =>
@@ -74,7 +77,8 @@ export const ButtonWrapper = styled.button<IButtonProps>`
   ${({ size }) =>
     size &&
     css`
-      padding: ${+calcSize(size, true) / 2}px ${calcSize(size)};
+      font-size: ${calcFontSize(size)};
+      padding: 0.5em 1em;
     `}
 
   ${({ primary, disabled }) =>
@@ -85,7 +89,7 @@ export const ButtonWrapper = styled.button<IButtonProps>`
       color: ${Colours.background};
 
       ${!disabled &&
-      `
+      css`
         &:hover {
           color: ${Colours.background};
           background-color: ${Colours.mainHover};
@@ -120,11 +124,6 @@ export const ButtonWrapper = styled.button<IButtonProps>`
       ${({ link }) =>
     link &&
     css`
-      & a {
-        color: ${Colours.main};
-        text-decoration: none;
-      }
-
       border: none;
       background: none;
       color: ${Colours.main};
@@ -140,64 +139,34 @@ export const ButtonWrapper = styled.button<IButtonProps>`
         color: ${Colours.mainActive};
       }
     `}
-
-    ${({ iconRight, iconLeft }) =>
-    (iconRight || iconLeft) &&
-    css`
-      display: flex;
-      align-items: center;
-    `}
 `;
 
-const Button = ({
-  className,
-  bordered,
-  isLoading,
-  onClick,
-  id,
-  name,
-  value,
-  text,
-  disabled,
-  type,
-  primary,
-  link,
-  size,
-  iconLeft: iconLeftDefault,
-  iconRight,
-  error,
-}: IButtonProps) => {
-  const iconLeft = isLoading ? <LoadingAnimation size="xs" inverted={primary || error} /> : iconLeftDefault;
+const Button = ({ className, isLoading, onClick, name, disabled, type, primary, link, size, icon: iconDefault, error, children, hasBorder }: IButtonProps) => {
+  const icon = isLoading ? <LoadingAnimation size="xs" inverted={primary || error} /> : iconDefault;
   return (
     <ButtonWrapper
       className={className}
-      id={id}
       name={name}
       type={type || 'button'}
-      value={value}
       disabled={disabled || isLoading}
       onClick={onClick}
       primary={primary}
       link={link}
       size={size}
-      iconLeft={iconLeft}
-      iconRight={iconRight}
+      icon={icon}
       data-test-id={`button-${name}`}
       error={error}
-      bordered={bordered}
+      hasBorder={hasBorder}
     >
-      {iconLeft}
-      <SText hasIconLeft={!!iconLeft} hasIconRight={!!iconRight}>
-        {text}
-      </SText>
-      {iconRight}
+      <span>{icon}</span>
+      <span>{children}</span>
     </ButtonWrapper>
   );
 };
 
 Button.defaultProps = {
   size: 'md',
-  bordered: true,
+  hasBorder: true,
   className: undefined,
 };